requirements.txt
fastapi
uvicorn  # async wsgi; waitress can be used as an option
cassandra-driver
python-dotenv
email-validator
argon2-cffi
pytest
jinja2
python-multipart
python-jose[cryptography]